Quick refactor to move json stuff into JsonLogAppender
